requirements.txt
Django==1.10.6
celery==4.0.2
requests==2.13.0
redis==2.10.5
PyYAML==3.12
docker==2.2.1
